modifs

+++ web/admin/gestion_liste/del_liste.php
@@ -8,20 +8,24 @@ require_once('../inc/droits.inc.php');
 
 <?php
 
-$id=$_GET['id'];
-$db = new DB_connection();
+if(isset($_GET['id']))
+{
+	$id=$_GET['id'];
+	$db = new DB_connection();
 
-$req="delete from compose where id_nivliste='".$id."'";
+	$req="delete from compose where id_nivliste='".$id."'";
 
-$db->DB_query($req);
+	$db->DB_query($req);
 
-$req="delete from inclus where id_nivliste='".$id."'";
-$db->DB_query($req);
+	$req="delete from inclus where id_nivliste='".$id."'";
+	$db->DB_query($req);
 
-$req="delete from liste_niveau where id_nivliste='".$id."'";
-$db->DB_query($req);
+	$req="delete from liste_niveau where id_nivliste='".$id."'";
+	$db->DB_query($req);
+
+	$db->DB_done();
+}
 $url="gestion_listes.php";
-
 header("Refresh:0;url=$url");
 
 ?>
